+++ scopes/component/component-drawer/index.ts
@@ -6,11 +6,12 @@ export {
   ComponentFiltersSlot,
 } from './component-drawer';
 export {
-  DrawerWidgetSlot,
-  ComponentTreeContextType,
-  ComponentTreeContext,
-  ComponentTreeProvider,
   ComponentFilterWidgetContextType,
   ComponentFilterWidgetContext,
   ComponentFilterWidgetProvider,
-} from './component-drawer.context';
+} from './component-drawer-filter-widget.context';
+export {
+  ComponentTreeContextType,
+  ComponentTreeContext,
+  ComponentTreeProvider,
+} from './component-drawer-tree-widget.context';
